moved section to seperate svelte files and added a second demo
