+++ app/src/App.svelte
@@ -1,8 +1,3 @@
-<!-- <script>
-	export let title;
-    gsap.to(".example", { duration: 1, x: 100 });
-
-</script> -->
 <div class="container">
 <header>
 			<h1>{title}</h1>
@@ -16,23 +11,9 @@
 				<acronym title="Greensock Animation Platform">GSAP</acronym> is the name
 				of the platform, which stands for <strong>G</strong>reen<strong>S</strong>ock <strong>A</strong>nimation <strong>P</strong>latform
 			</p>
-<section>
-<h2>A first simple animation</h2>
-<p> Here come a first example:</p>
-<div class="demo__container" id="simpleDemo">
-<h2 class="demo__title">A simple demo: <code>gsap.to</code></h2>
-{#if !example1Started}
-<button class="button start" on:click={startAnimation} >
-	Start &RightAngleBracket;
-</button>
-{:else}
-<button class="button reset" on:click={resetAnimation} >
-	Reset &ldca;
-</button>
-{/if}
-<div class="example"></div>
-</div>
-</section>
+
+<Demo1/>
+<Demo2/>
 		</main>
         <aside>
         <ul>
@@ -40,18 +21,7 @@
             <li class="link"><a href="#simpleDemo">Simple Demo</a></li>
         </ul>
         </aside>
-<footer>
-<p class="footer__content">2020 - Personal project by Johan Bijlsma
-</p>
-<div class="footer__socials">
-<a href="https://twitter.com/johanbijlsma"> <img src="/icon-social-twitter-white.svg" alt="Twitter Logo"/>
-<span class="footer__socials--text"> Johan's Twitter </span></a>
-<a href="https://github.com/johanbijlsma"> <img src="/Github.svg" alt="Github Logo"/>
-<span class="footer__socials--text"> Johan's Github </span></a>
-<a href="https://codepen.io/johanbijlsma/"> <img src="/codepen.svg" alt="codepen Logo"/>
-<span class="footer__socials--text"> Johan's Codepen </span></a>
-</div>
-</footer>
+<Footer/>
 
 
 </div>
@@ -102,52 +72,7 @@
         background-color: #8AC640;
         columns: #333333;
     }
-    footer{
-        grid-column-start: 1;
-        grid-column-end: 6;
-        grid-row-start: 3;
-        grid-row-end: 3;
-    }
-    .footer__content{
-          grid-column-start: 2;
-          text-align: center;
-    }
-    .footer__socials{
-        display: flex;
-        justify-content: space-evenly;
-        max-width: 60vw;
-        margin: 10px auto;
-        }
-    .footer__socials a img{
-        height: 30px;
-        transition: all 800ms 30ms cubic-bezier(0.165, 0.84, 0.44, 1);
-    }
-    .footer__socials a{
-        height: 30px;
-        width: fit-content;
-        padding: 10px;
-        transition: all 800ms 30ms cubic-bezier(0.165, 0.84, 0.44, 1);
-    }
 
-    .footer__socials a:hover, .footer__socials a:focus{
-        background: lightslategrey;
-        padding: 10px 30px;
-        transition: all 800ms 120ms ease;
-        font-weight: 700;
-        padding-bottom: 40px;
-        box-shadow: 0 3px 0 #8AC640;
-    }
-    .footer__socials a:hover img, .footer__socials a:focus img{
-        transform: scale(1.25);
-        padding-right: 12px;
-        transition: all 800ms 30ms cubic-bezier(0.165, 0.84, 0.44, 1);
-
-    }
-    .footer__socials a span.footer__socials--text{
-        transform: translate(0px, -10px);
-        display: inline-block;
-        color: #8AC640;
-    }
 
 	@media (min-width: 640px) {
 		main {
@@ -165,77 +90,24 @@
         src: url("/Jost-VariableFont_ital,wght.ttf");
     }
 
-    header, footer {
+    header{
     color: #8AC640;
 
         height: 120px;
         background: #333333;
     }
-    .demo__container{
-        margin: 10px auto;
-        width: 80%;
-        border: dashed 2px rgba(0,0,30,0.3);
-        padding: 20px;
-        background-color: #33333375;
-        border-radius: 30px;
-    }
 
-    .demo__title{
-            color: #8AC640;
 
-    }
 
-    .example{
-        width: 35px;
-        height: 35px;
-        background-color: red;
-        border-radius: 10px;
-        border: 2px solid white;
-    }
-
-code{
-    font-family: monospace;
-    color: #8AC640;
-}
-
-button.button{
-    border-radius: 10px;
-    border: 1px solid darkblue;
-    padding: 8px 32px;
-    box-shadow: inset 0 0 0 transparent;
-}
-.button.start:hover{
-    box-shadow: inset 0 -6px 1px #8AC640;
-    cursor: pointer;
-    transition: all 200ms ease-in-out;
-    color: #8AC640;
-}
-
-.button.reset{
-    box-shadow: inset 0 -40px 1px rgb(204, 63, 8);
-    transition: all 200ms ease-in-out;
-    color: #fff;
-    }
-
-    .button.reset:hover{
-    box-shadow: inset 0 -6px 1px #8AC640;
-    cursor: pointer;
-    transition: all 200ms ease-in-out;
-    color: #8AC640;
-}
 </style>
 
 <script>
+
+import  Footer  from "./Footer.svelte";
+import  Demo1  from "./Demo1.svelte";
+import  Demo2  from "./Demo2.svelte";
 	export let title;
-    export let example1Started = false;
-	function startAnimation(event) {
-    gsap.to('.example', { duration: 1, x: 100 });
-    example1Started = true;
-	}
-	function resetAnimation(event) {
-    gsap.to('.example', { duration: 1, x: 0 });
-    example1Started = false;
-	}
+
 
     // console.log(container);
