+++ app/src/main.js
@@ -1,4 +1,5 @@
 import App from "./App.svelte";
+
 // import { gsap } from "gsap";
 
 const app = new App({
@@ -6,12 +7,6 @@ const app = new App({
   props: {
     title: `Learning Greensock's GSAP`,
   },
-  //  GS = gsap
 });
-
-// const GS = gsap({
-//     GS(to(".example1", { duration: 1, x: 100 }));
-
-// })
 
 export default app;
